Moved JSON storage to delivery.webserver.cache
