+++ Plan/common/src/main/java/com/djrapitops/plan/delivery/webserver/cache/JSONStorage.java
@@ -14,7 +14,7 @@
  *  You should have received a copy of the GNU Lesser General Public License
  *  along with Plan. If not, see <https://www.gnu.org/licenses/>.
  */
-package com.djrapitops.plan.storage.json;
+package com.djrapitops.plan.delivery.webserver.cache;
 
 import com.djrapitops.plan.SubSystem;
 import com.google.gson.Gson;
