+++ Plan/common/src/main/java/com/djrapitops/plan/modules/SystemObjectProvidingModule.java
@@ -19,15 +19,15 @@ package com.djrapitops.plan.modules;
 import com.djrapitops.plan.DataService;
 import com.djrapitops.plan.DataSvc;
 import com.djrapitops.plan.PlanPlugin;
+import com.djrapitops.plan.delivery.webserver.cache.JSONFileStorage;
+import com.djrapitops.plan.delivery.webserver.cache.JSONMemoryStorageShim;
+import com.djrapitops.plan.delivery.webserver.cache.JSONStorage;
 import com.djrapitops.plan.gathering.importing.importers.Importer;
 import com.djrapitops.plan.settings.config.ExtensionSettings;
 import com.djrapitops.plan.settings.config.PlanConfig;
 import com.djrapitops.plan.settings.locale.Locale;
 import com.djrapitops.plan.settings.locale.LocaleSystem;
 import com.djrapitops.plan.storage.file.JarResource;
-import com.djrapitops.plan.storage.json.JSONFileStorage;
-import com.djrapitops.plan.storage.json.JSONMemoryStorageShim;
-import com.djrapitops.plan.storage.json.JSONStorage;
 import com.djrapitops.plan.utilities.logging.ErrorLogger;
 import com.djrapitops.plan.utilities.logging.PluginErrorLogger;
 import dagger.Module;
